Pro-Coder/stock_ai_boat_web/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
pillow
python-dotenv
sqlmodel
passlib[bcrypt]
pyjwt
scikit-learn
tensorflow>=2.11
